move config up a folder
