+++ src/cli/routes.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const CoreConfig = require("../core/config/index.js");
+const CoreConfig = require("../core/Config");
 const fs = require("fs");
 
 const NounHelper = require("../core/helpers/NounHelper");
